Bug 973445 - [settings] refactor Sound panel with AMD pattern
