+++ apps/settings/elements/sound.html
@@ -106,8 +106,7 @@
       </ul>
     </div>
 
-    <script src="shared/js/omadrm/fl.js"></script>
-    <script src="js/sound.js"></script>
+    <panel data-path="panels/sound/panel"></panel>
 
   </template>
 </element>
